+++ app/src/main/java/org/simple/clinic/facility/change/confirm/FacilityChangeComponent.kt
@@ -1,7 +1,8 @@
 package org.simple.clinic.facility.change.confirm
 
+import androidx.appcompat.app.AppCompatActivity
+import dagger.BindsInstance
 import dagger.Subcomponent
-import org.simple.clinic.activity.BindsActivity
 import org.simple.clinic.di.AssistedInjectModule
 import org.simple.clinic.facility.change.FacilityChangeActivity
 import org.simple.clinic.facilitypicker.FacilityPickerView
@@ -9,13 +10,12 @@ import org.simple.clinic.facilitypicker.FacilityPickerView
 @Subcomponent(
     modules = [AssistedInjectModule::class]
 )
-interface FacilityChangeComponent: FacilityPickerView.Injector {
+interface FacilityChangeComponent : FacilityPickerView.Injector {
 
   fun inject(activity: FacilityChangeActivity)
 
-  @Subcomponent.Builder
-  interface Builder : BindsActivity<Builder> {
-
-    fun build(): FacilityChangeComponent
+  @Subcomponent.Factory
+  interface Factory {
+    fun create(@BindsInstance activity: AppCompatActivity): FacilityChangeComponent
   }
 }