+++ app/src/main/java/org/simple/clinic/scheduleappointment/di/ScheduleAppointmentSheetComponent.kt
@@ -1,7 +1,8 @@
 package org.simple.clinic.scheduleappointment.di
 
+import androidx.appcompat.app.AppCompatActivity
+import dagger.BindsInstance
 import dagger.Subcomponent
-import org.simple.clinic.activity.BindsActivity
 import org.simple.clinic.di.AssistedInjectModule
 import org.simple.clinic.scheduleappointment.ScheduleAppointmentSheet
 
@@ -10,8 +11,8 @@ interface ScheduleAppointmentSheetComponent {
 
   fun inject(target: ScheduleAppointmentSheet): ScheduleAppointmentSheet
 
-  @Subcomponent.Builder
-  interface Builder: BindsActivity<Builder> {
-    fun build(): ScheduleAppointmentSheetComponent
+  @Subcomponent.Factory
+  interface Factory {
+    fun create(@BindsInstance activity: AppCompatActivity): ScheduleAppointmentSheetComponent
   }
 }