+++ app/src/main/java/org/simple/clinic/teleconsultlog/medicinefrequency/di/MedicineFrequencyComponent.kt
@@ -1,7 +1,8 @@
 package org.simple.clinic.teleconsultlog.medicinefrequency.di
 
+import androidx.appcompat.app.AppCompatActivity
+import dagger.BindsInstance
 import dagger.Subcomponent
-import org.simple.clinic.activity.BindsActivity
 import org.simple.clinic.di.AssistedInjectModule
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequencySheet
 
@@ -10,8 +11,8 @@ interface MedicineFrequencyComponent {
 
   fun inject(target: MedicineFrequencySheet)
 
-  @Subcomponent.Builder
-  interface Builder : BindsActivity<Builder> {
-    fun build(): MedicineFrequencyComponent
+  @Subcomponent.Factory
+  interface Factory {
+    fun create(@BindsInstance activity: AppCompatActivity): MedicineFrequencyComponent
   }
 }